Moved query.run_query to daac_data_subscriber.py to break circular import dependencies
